feat: SSL support or Websockets
