+++ conf/mock/node.properties
@@ -1,0 +1,8 @@
+node.network = signum.net.MockNetwork
+DB.SqliteJournalMode = WAL
+
+# Database connection JDBC url
+DB.Url=jdbc:sqlite:file:./db/signum-test.sqlite.db
+
+SoloMiningPassphrases="some phrase"
+AllowOtherSoloMiners=true
