+++ conf/mock/node.sqlite-mock.properties
@@ -1,0 +1,4 @@
+DB.SqliteJournalMode = WAL
+
+# Database connection JDBC url
+DB.Url=jdbc:sqlite:file:./db/signum-mock.sqlite.db
